move AddWorkerNodes out of button directory
