+++ src/components/CostWizard/WizardSteps/WorkerNodesStep.tsx
@@ -2,11 +2,9 @@ import React from 'react';
 import { Title, WizardStep } from '@ui5/webcomponents-react';
 import NextStepButton from '../Buttons/NextStepButton';
 import PreviousStepButton from '../Buttons/PreviousStepButton';
-import AddWorkerNodes from '../Buttons/AddWorkerNodes';
+import AddWorkerNodes from '../AddWorkerNodes';
 
 export default function WorkerNodesStep() {
-
-
   return (
     <WizardStep disabled titleText="Worker Node Pools">
       <Title wrappingType="Normal" level="H2" size="H2">
